+++ fabrictestbed_extensions/fablib/fablib.py
@@ -33,6 +33,7 @@ from ipaddress import IPv4Network, IPv6Network
 from typing import TYPE_CHECKING, Dict, List
 
 import pandas as pd
+import paramiko
 from fabrictestbed.util.constants import Constants
 from IPython import get_ipython
 from IPython.core.display_functions import display
@@ -608,6 +609,7 @@ class FablibManager:
         data_dir: str = None,
         output: str = None,
         execute_thread_pool_size: int = 64,
+        offline: bool = False,
     ):
         """
         Constructor. Builds FablibManager.  Tries to get configuration from:
@@ -789,7 +791,9 @@ class FablibManager:
         self.resources = None
         self.links = None
         self.facility_ports = None
-        self.build_slice_manager()
+
+        if not offline:
+            self.build_slice_manager()
 
     def _validate_configuration(self):
         """
@@ -1598,6 +1602,16 @@ class FablibManager:
         """
         return self.bastion_key_filename
 
+    def get_bastion_key(self) -> str:
+        """
+        Reads the FABRIC Bastion private key file and returns the key.
+
+        :return: FABRIC Bastion key string
+        :rtype: String
+        """
+        with open(self.bastion_key_filename, "r", encoding="utf-8") as f:
+            return f.read()
+
     def get_bastion_public_addr(self) -> str:
         """
         Gets the FABRIC Bastion host address.
@@ -1626,6 +1640,57 @@ class FablibManager:
         :rtype: String
         """
         return self.bastion_private_ipv6_addr
+
+    def probe_bastion_host(self) -> bool:
+        """
+        See if bastion will admit us with our configuration.
+
+        Bastion hosts are configured to block hosts that attempts to
+        use it with too many repeated authentication failures.  We
+        want to avoid that.
+
+        Returns ``True`` if connection attempt succeeds.  Raises an
+        error in the event of failure.
+        """
+
+        bastion_client = paramiko.SSHClient()
+        bastion_client.set_missing_host_key_policy(paramiko.AutoAddPolicy)
+
+        try:
+            bastion_host = self.get_bastion_public_addr()
+            bastion_username = self.get_bastion_username()
+            bastion_key_path = self.get_bastion_key_filename()
+            bastion_key_passphrase = self.bastion_passphrase
+
+            logging.info(
+                f"Probing bastion host {bastion_host} with "
+                f"username: {bastion_username}, key: {bastion_key_path}, "
+                f"key passphrase: {'hidden' if bastion_key_passphrase else None}"
+            )
+
+            result = bastion_client.connect(
+                hostname=bastion_host,
+                username=bastion_username,
+                key_filename=bastion_key_path,
+                passphrase=bastion_key_passphrase,
+                look_for_keys=False,
+            )
+
+            # Things should be fine if we are here.
+            if result is None:
+                logging.info(f"Connection with {bastion_host} appears to be working")
+                return True
+
+        # In theory paramiko can raise several types of exceptions,
+        # but in practice it has not been that precise.  Let us treat
+        # all exceptions as un-recoverable for now, and refine the
+        # behavior based on some real-world testing.
+        except (paramiko.SSHException, Exception) as e:
+            logging.error(f"Bastion connection error: {e}")
+            raise e
+
+        finally:
+            bastion_client.close()
 
     def set_slice_manager(self, slice_manager: SliceManager):
         """
@@ -1932,7 +1997,12 @@ class FablibManager:
                 excludes=[SliceState.Dead, SliceState.Closing], slice_name=name
             )
 
-            return slices[0]
+            if len(slices) > 0:
+                return slices[0]
+            else:
+                raise Exception(
+                    f'Unable to find slice "{name}" for this project. Check slice name spelling and project id.'
+                )
         else:
             raise Exception(
                 "get_slice requires slice name (name) or slice id (slice_id)"