+++ tests/integration/test_list_resources.py
@@ -1,0 +1,47 @@
+#!/usr/bin/env python3
+# MIT License
+#
+# Copyright (c) 2020 FABRIC Testbed
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+# Author Komal Thareja (kthare10@renci.org)
+import unittest
+
+from fabrictestbed_extensions.fablib.fablib import FablibManager
+
+
+class ListResourcesTests(unittest.TestCase):
+    def test_list_resources(self):
+        # Import the FABlib Library
+        fablib = FablibManager()
+
+        fablib.show_config()
+
+        print("List Sites")
+        sites = fablib.list_sites(output="json")
+        self.assertIsNotNone(sites)
+
+        print("List Facilities")
+        facilities = fablib.list_facility_ports(output="json")
+        self.assertIsNotNone(facilities)
+
+
+if __name__ == "__main__":
+    unittest.main()