+++ CHANGELOG.md
@@ -8,17 +8,42 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
 - Address a crash when querying NUMA properties. (Issue
   [#191](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/191),
   PR [#192](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/192))
+
+### Changed
+
+- Update list of OS images (PR
+  [#202](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/202))
+- List Facility Ports updated to include additional parameters (Issue
+  [#210](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/210))
+- Fail early when connection with bastion host fails (Issue
+  [#151](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/151))
+- Overhaul API docs (PR
+  [#217](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/217),
+  issue [#215](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/215))
+
+### Added
+
 - Add/update integration tests (Issues
   [#184](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/184),
   [#186](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/186),
   PR [#187](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/187))
+- Make Network Interface Config Idempotant
+  (Issue [#205](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/205))
+- Methods added to retrieve SSH keys for bastion and sliver (PR
+  [#207](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/207))
+- Support for PortMirror service (PR
+  [#214](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/214))
+
+### Removed
+  
 - Remove unused `AbcFabLIB` class (Issue
   [#117](https://github.com/fabric-testbed/fabrictestbed-extensions/issues/117))
-- Update list of OS images (PR
-  [#202](https://github.com/fabric-testbed/fabrictestbed-extensions/pull/202))
+
 
 ## [1.4.4] - 2023-05-21